refactor(handlers): Extract calendar API handler
Move the `apiCalendar` handler and its associated types to a new
`api_calendar_handlers.go` file. This refactoring improves the separation of
concerns by dedicating a file to calendar-related API logic, reducing the size
and complexity of `api_handlers.go`, and making the codebase easier to navigate
and maintain.

Introduce `calendarQueryParams` to encapsulate the logic for handling `start`
and `end` query parameters. This change centralizes date parsing and database
filtering for calendar queries, making the handler cleaner and more robust by
explicitly handling potential parsing errors for date parameters.

Define `calTS` (calendar timestamp) constant for the expected date format. This
constant makes the date format explicit and ensures consistency across date
parsing operations related to the calendar, improving readability and
maintainability.

Fixes #577
